requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
